update route for RBAC

+++ app/api/categories/[id]/route.ts
@@ -1,5 +1,7 @@
 import { deleteCategoryController, getSingleCategoryController, updateCategoryController } from "@/backend/controllers/categoryController";
-import { NextResponse } from "next/server";
+import { verifyToken } from "@/backend/middleware/authMiddleware";
+import { cookies } from "next/headers";
+import { NextRequest, NextResponse } from "next/server";
 
 export async function GET(request: NextResponse, { params }: { params: Promise<{ id: string }> }){
     try {
@@ -13,27 +15,56 @@ export async function GET(request: NextResponse, { params }: { params: Promise<{
     }
 }
 
-export async function PUT(request: NextResponse, { params }: { params: Promise<{ id: string }> }){
+export async function PUT( request: NextRequest, { params }: { params: Promise<{ id: string }> }) {
     try {
-        const {id} = await params
+        const cookieStore = await cookies();
 
-        const body = await request.json()
+        const token = cookieStore.get("token")?.value;
 
-        const result = await updateCategoryController(Number(id), body)
+        if (!token) {
+            return NextResponse.json( { message: "Unauthorized" },{ status: 401 });
+        }
 
-        return NextResponse.json(result)
+        const user = verifyToken(token);
+
+        if (user.role !== "ADMIN") {
+            return NextResponse.json( { message: "Access Denied. Admin only." }, { status: 403 });
+        }
+
+        const { id } = await params;
+
+        const body = await request.json();
+
+        const result = await updateCategoryController( Number(id), body);
+
+        return NextResponse.json(result);
     } catch (error) {
-        return NextResponse.json( { message: "Error" }, { status: 500 } );
+        return NextResponse.json( { message: "Error" }, { status: 500 });
     }
-}
+    }
 
-export async function DELETE(request:NextResponse, { params }: { params: Promise<{ id: string }> }){
+export async function DELETE(request: NextRequest, { params }: { params: Promise<{ id: string }> }) {
     try {
-        const {id} = await params
+        const cookieStore = await cookies();
 
-        const result = await deleteCategoryController(Number(id))
-        return NextResponse.json(result)
+        const token = cookieStore.get("token")?.value;
+
+        if (!token) {
+            return NextResponse.json({ message: "Unauthorized" }, { status: 401 } );
+        }
+
+        const user = verifyToken(token);
+
+        if (user.role !== "ADMIN") {
+            return NextResponse.json( { message: "Access Denied. Admin only." }, { status: 403 } );
+        }
+
+        const { id } = await params;
+
+        const result = await deleteCategoryController( Number(id) );
+
+        return NextResponse.json(result);
     } catch (error) {
-        return NextResponse.json({ message: "Error" }, { status: 500 })
+        return NextResponse.json( { message: "Error" }, { status: 500 });
     }
 }
